// src/lib/api-types.ts
export interface ImageConfig {
  url: string;
  title: string;
  link?: string;
}

export interface CustomerServiceConfig {
  img: string;
  qq: string;
}

export interface SiteConfig {
  id: number;
  carousel: ImageConfig[];
  bannerL: ImageConfig;
  bannerR: ImageConfig;
  customerService: CustomerServiceConfig;
  updated_at: string;
}